feat(sessions): drop the unmerged-work kill guard (#1579)
`af sessions kill` (and the TUI `D` action and daemon KillSession RPC) shared
a fail-safe guard that refused to destroy a session whose branch had
commits-not-on-base, a dirty worktree, or a branch mismatch, unless `--force`
was passed. In practice it over-refused ordinary cleanup:

- Squash-merged branches: after the default `gh pr merge --squash`, the
  branch's original commits are no longer ancestors of the base even though
  every change is fully landed, so the guard reported "N commits not on
  master" and blocked the kill.
- Branch mismatch: when a worktree was checked out on a different branch than
  the stored session branch, the guard failed with "could not verify safe to
  destroy: worktree is on branch X but the stored session branch is Y".

Either case blocked routine session deletion — roughly half of a heavy user's
sessions became un-killable without `--force`. Per owner decision, drop the
guard entirely: kill now always destroys the session. `af sessions archive`
remains the non-destructive, restorable default. `--force` is retained as an
accepted no-op so existing `af sessions kill --force` invocations keep working.

- Delete daemon/kill_guard.go and its whole helper set.
- Remove the guard call from daemon.KillSession; document Force as a no-op.
- Update `af sessions kill` help + docs/cli.md + regenerated CLI reference.
- Replace the guard-refusal tests with tests proving kill destroys unmerged,
  dirty, and branch-mismatched sessions WITHOUT --force, and that --force is
  still accepted.

Fixes #1579
