+++ api/sessions.go
@@ -735,9 +735,9 @@ Factory owns it.
 For normal "done with this session" cleanup, prefer:
   af sessions archive <title>
 
-Kill refuses by default unless it can confirm the worktree is clean and HEAD
-has no commits unreachable from the session's base/default branch. Use --force
-to skip that guard and destroy the session anyway.`,
+Kill always destroys the session, including any uncommitted or unmerged work on
+its branch — there is no undo. To keep a session restorable instead, archive it.
+--force is accepted but has no effect (kept for backward compatibility).`,
 	Args: cobra.ExactArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 		log.Initialize(false)
